+++ tests/test_extract.py
@@ -361,6 +361,19 @@ class TestReuseInfoOfFile:
             assert result == ReuseInfo()
         assert f"'{path}' was detected as a binary file" in caplog.text
 
+    def test_log(self, caplog):
+        """Log the extraction to with level logging.DEBUG"""
+        caplog.set_level(logging.DEBUG)
+        buffer = BytesIO(b"# Copyright Jane Doe")
+        buffer.name = "foo.py"
+        reuse_info_of_file(buffer)
+        assert len(caplog.records) == 1
+        assert caplog.records[0].levelname == "DEBUG"
+        assert caplog.records[0].msg == (
+            "extracting REUSE information from 'foo.py'"
+            " (encoding 'utf_8', newline '\\n')"
+        )
+
     @pytest.mark.parametrize("newline", ["\r\n", "\r", "\n"])
     def test_all_newlines(self, newline):
         """Can lint files with any newline."""