+++ CHANGELOG.md
@@ -55,6 +55,7 @@ You will not encounter this bug if your environment has libmagic available.
 - The Docker images and the pre-commit hooks now come bundled with all encoding
   modules. (#1245)
 - The `--debug` flag now tells you which encoding module is used. (#1246)
+- The `--debug` flag now tells you the detected encoding of each file. (#1246)
 
 ## v6.0.0 - 2025-10-06